+++ cogs/misc.py
@@ -9,20 +9,24 @@ class Misc:
     def __init__(self, bot):
         self.bot = bot
 
+
     @commands.command(pass_context=True)
-    async def feedback(self, ctx):
+    async def feedback(self, ctx, *args):
         """Allow user to send feedback to the bot's developer"""
         user = ctx.message.author
-        manager = MessageManager(self.bot, user, ctx.message.channel, ctx.message)
+        manager = MessageManager(self.bot, user, ctx.message.channel, [ctx.message])
 
-        res = await manager.say_and_wait("Enter the feedback you would like to send to the developer")
-        if not res:
-            return
-        feedback = res.content
-        await manager.say("Feedback received! Thank you for your input.")
+        if len(args) == 0:
+            await manager.say("You forgot to include your feedback!")
+            return await manager.clear()
+
+        feedback = ""
+        for word in args:
+            feedback += "{} ".format(word)
 
         # Send user feedback to Asal, the bot's devloper
         asal = await self.bot.get_user_info("118926942404608003")
         await self.bot.send_message(asal, "Feedback from {}:\n---\n{}".format(user.mention, feedback))
 
+        await manager.say("Your feedback has been sent to the devloper. Thank you for your input!")
         await manager.clear()