+++ cogs/settings.py
@@ -13,32 +13,30 @@ class Settings:
     def __init__(self, bot):
         self.bot = bot
 
+
     @commands.command(pass_context=True)
     async def setprefix(self, ctx, new_prefix=None):
-
+        """Set a custom server prefix"""
         user = ctx.message.author
-        manager = MessageManager(self.bot, user, ctx.message.channel, ctx.message)
+        channel = ctx.message.channel
+        manager = MessageManager(self.bot, user, channel, [ctx.message])
 
-        server = None
         if ctx.message.channel.is_private:
-            server = get_server_from_dm(self.bot, ctx)
-            if not server:
-                return await manager.say("You must be part of only one {} server to do that in a DM".format(self.bot.user.mention))
-        else:
-            server = ctx.message.server
+            return await manager.say("That command is not supported in a direct message.")
 
-        if not is_admin(user, server):
+        if not is_admin(user, channel):
             await manager.say("You must be an admin to do that.")
             return await manager.clear()
 
         if not new_prefix:
-            await manager.say("You must provide a new prefix.")
+            await manager.say("You must provide a prefix.")
+            return await manager.clear()
 
         if len(new_prefix) > 5:
-            await manager.say("Oops! Prefix must be less than 6 characters.")
+            await manager.say("Prefix must be less than 6 characters.")
             return await manager.clear()
 
         with DBase() as db:
-            db.set_prefix(server.id, new_prefix)
+            db.set_prefix(ctx.message.server.id, new_prefix)
             await manager.say("Command prefix has been changed to " + new_prefix)
             return await manager.clear()